third_party/blink/web_tests/platform/generic/fast/forms/select/select-element-item-argument-expected.txt

Tests that the HTMLSelectElement.item() argument is correctly validated.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS select.__proto__ is HTMLSelectElement.prototype
PASS select.item(0).value is "a"
PASS select.item(1).value is "b"
PASS select.item(2).value is "c"
PASS select.item(3).value is "d"
PASS select.item(4) is null
PASS select.item(-1) is null
PASS select.item(-4294967294).value is "c"
PASS select.item() threw exception TypeError: Failed to execute 'item' on 'HTMLSelectElement': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE

a
b
c
d
